refactor: extract structure tree operations
